+++ attributes/default.rb
@@ -1,2 +1,3 @@
-default['domain'] = "priv.evertrue.com"
-default['fqdn'] = "#{node.name}.#{node['domain']}"
+set['target_domain'] = "priv.evertrue.com"
+set['target_fqdn'] = "#{node.name}.#{node['target_domain']}"
+override['domain'] = node['target_domain']