+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of hostname.
 
+## 1.0.2:
+
+* Maintain /etc/hostname file
+* Refactor /etc/hosts template
+
 ## 1.0.1:
 
 * Total refactor